+++ changelog.md
@@ -5,6 +5,13 @@ This change log records updates after 2025-3-24.
 
 Please view [SomeB1oody/RustyML](https://github.com/SomeB1oody/RustyML) for more info.
 
+## [v0.6.0] - 2025-5-3
+### Changed
+- Refactor parameter updates to use parallel processing
+
+### Removed
+- Remove method `update_parameters` in `Layer` trait
+
 ## [v0.6.0] - 2025-5-2
 ### Changed
 - Refactor RMSprop implementation with unified cache structure